requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
dicom==0.9.9.post1
exceptiongroup==1.2.0
executing==2.0.1
fast-histogram==0.12
fonttools==4.50.0
gdcm==1.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipykernel==6.29.3
ipympl==0.9.3
ipython==8.22.2
ipython-genutils==0.2.0
ipywidgets==8.1.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
moviepy==1.0.3
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydicom==2.4.4
pygame==2.5.2
Pygments==2.17.2
pyjpegls==1.3.0
pylibjpeg==2.0.0
pylibjpeg-libjpeg==2.0.2
pylibjpeg-openjpeg==2.1.1
pylibjpeg-rle==2.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.23
pywin32==306
pyzmq==25.1.2
requests==2.31.0
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
urllib3==2.2.1
wcwidth==0.2.13
widgetsnbextension==4.0.10
